Split to different files and fixed comments.
