+++ librz/util/thread_lock.c
@@ -1,50 +1,46 @@
 // SPDX-FileCopyrightText: 2009-2017 pancake <pancake@nopcode.org>
+// SPDX-FileCopyrightText: 2022 deroad <wargio@libero.it>
 // SPDX-License-Identifier: LGPL-3.0-only
 
 #include "thread.h"
 
-/* locks/mutex/sems */
-
-RZ_API RzThreadLock *rz_th_lock_new(bool recursive) {
+RZ_API RZ_OWN RzThreadLock *rz_th_lock_new(bool recursive) {
 	RzThreadLock *thl = RZ_NEW0(RzThreadLock);
-	if (thl) {
-#if HAVE_PTHREAD
-		if (recursive) {
-			pthread_mutexattr_t attr;
-			pthread_mutexattr_init(&attr);
-#if !defined(__GLIBC__) || __USE_UNIX98__
-			pthread_mutexattr_settype(&attr, PTHREAD_MUTEX_RECURSIVE);
-#else
-			pthread_mutexattr_settype(&attr, PTHREAD_MUTEX_RECURSIVE_NP);
-#endif
-			pthread_mutex_init(&thl->lock, &attr);
-		} else {
-			pthread_mutex_init(&thl->lock, NULL);
-		}
-#elif __WINDOWS__
-		// TODO: obey `recursive` (currently it is always recursive)
-		InitializeCriticalSection(&thl->lock);
-#endif
+	if (!thl) {
+		return NULL;
 	}
+#if HAVE_PTHREAD
+	if (recursive) {
+		pthread_mutexattr_t attr;
+		pthread_mutexattr_init(&attr);
+#if !defined(__GLIBC__) || __USE_UNIX98__
+		pthread_mutexattr_settype(&attr, PTHREAD_MUTEX_RECURSIVE);
+#else
+		pthread_mutexattr_settype(&attr, PTHREAD_MUTEX_RECURSIVE_NP);
+#endif
+		pthread_mutex_init(&thl->lock, &attr);
+	} else {
+		pthread_mutex_init(&thl->lock, NULL);
+	}
+#elif __WINDOWS__
+	// Windows critical sections always accept recursive
+	// access and it cannot be configured in any other way.
+	InitializeCriticalSection(&thl->lock);
+#endif
 	return thl;
 }
 
-RZ_API int rz_th_lock_wait(RzThreadLock *thl) {
-	rz_th_lock_enter(thl); // locks here
-	rz_th_lock_leave(thl); // releases previous mutex
-	return 0;
-}
-
-RZ_API int rz_th_lock_enter(RzThreadLock *thl) {
+RZ_API void rz_th_lock_enter(RZ_NONNULL RzThreadLock *thl) {
+	rz_return_if_fail(thl);
 #if HAVE_PTHREAD
-	return pthread_mutex_lock(&thl->lock);
+	pthread_mutex_lock(&thl->lock);
 #elif __WINDOWS__
 	EnterCriticalSection(&thl->lock);
-	return 0;
 #endif
 }
 
-RZ_API int rz_th_lock_tryenter(RzThreadLock *thl) {
+RZ_API bool rz_th_lock_tryenter(RZ_NONNULL RzThreadLock *thl) {
+	rz_return_val_if_fail(thl, false);
 #if HAVE_PTHREAD
 	return !pthread_mutex_trylock(&thl->lock);
 #elif __WINDOWS__
@@ -52,23 +48,23 @@ RZ_API int rz_th_lock_tryenter(RzThreadLock *thl) {
 #endif
 }
 
-RZ_API int rz_th_lock_leave(RzThreadLock *thl) {
+RZ_API void rz_th_lock_leave(RZ_NONNULL RzThreadLock *thl) {
+	rz_return_if_fail(thl);
 #if HAVE_PTHREAD
-	return pthread_mutex_unlock(&thl->lock);
+	pthread_mutex_unlock(&thl->lock);
 #elif __WINDOWS__
 	LeaveCriticalSection(&thl->lock);
-	return 0;
 #endif
 }
 
-RZ_API void *rz_th_lock_free(RzThreadLock *thl) {
-	if (thl) {
-#if HAVE_PTHREAD
-		pthread_mutex_destroy(&thl->lock);
-#elif __WINDOWS__
-		DeleteCriticalSection(&thl->lock);
-#endif
-		free(thl);
+RZ_API void rz_th_lock_free(RZ_NULLABLE RzThreadLock *thl) {
+	if (!thl) {
+		return;
 	}
-	return NULL;
+#if HAVE_PTHREAD
+	pthread_mutex_destroy(&thl->lock);
+#elif __WINDOWS__
+	DeleteCriticalSection(&thl->lock);
+#endif
+	free(thl);
 }
