+++ librz/util/thread_sem.c
@@ -1,4 +1,5 @@
 // SPDX-FileCopyrightText: 2018 thestr4ng3r <info@florianmaerkl.de>
+// SPDX-FileCopyrightText: 2022 deroad <wargio@libero.it>
 // SPDX-License-Identifier: LGPL-3.0-only
 
 #include "thread.h"
@@ -15,7 +16,7 @@
 #include <limits.h>
 #endif
 
-RZ_API RzThreadSemaphore *rz_th_sem_new(unsigned int initial) {
+RZ_API RZ_OWN RzThreadSemaphore *rz_th_sem_new(unsigned int initial) {
 	RzThreadSemaphore *sem = RZ_NEW(RzThreadSemaphore);
 	if (!sem) {
 		return NULL;
@@ -57,7 +58,7 @@ RZ_API RzThreadSemaphore *rz_th_sem_new(unsigned int initial) {
 	return sem;
 }
 
-RZ_API void rz_th_sem_free(RzThreadSemaphore *sem) {
+RZ_API void rz_th_sem_free(RZ_NULLABLE RzThreadSemaphore *sem) {
 	if (!sem) {
 		return;
 	}
@@ -76,7 +77,8 @@ RZ_API void rz_th_sem_free(RzThreadSemaphore *sem) {
 	free(sem);
 }
 
-RZ_API void rz_th_sem_post(RzThreadSemaphore *sem) {
+RZ_API void rz_th_sem_post(RZ_NONNULL RzThreadSemaphore *sem) {
+	rz_return_if_fail(sem);
 #if HAVE_PTHREAD
 	sem_post(sem->sem);
 #elif __WINDOWS__
@@ -84,7 +86,8 @@ RZ_API void rz_th_sem_post(RzThreadSemaphore *sem) {
 #endif
 }
 
-RZ_API void rz_th_sem_wait(RzThreadSemaphore *sem) {
+RZ_API void rz_th_sem_wait(RZ_NONNULL RzThreadSemaphore *sem) {
+	rz_return_if_fail(sem);
 #if HAVE_PTHREAD
 	sem_wait(sem->sem);
 #elif __WINDOWS__
